+++ library/sdl.cpp
@@ -73,7 +73,7 @@ namespace sdl {
   SDL_Delay(ms);
  }
 
- u32 get_ticks() {
+ u64 get_ticks() {
   return SDL_GetTicks();
  }
